Consolidate basic operations
